backend/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-CORS==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
pandas==2.1.1
openpyxl==3.1.2
python-dateutil==2.8.2